requirements.txt
streamlit
langchain-community
langchain-groq
langchain-classic
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
sentence-transformers
torch
torchvision